Add files via upload
Discover the inner workings of bubble sort, merge sort, linear search, and binary search with this Python GUI project. Built with Tkinter, this program generates a randomized array of numbers for visualizing these algorithms step-by-step. With adjustable animation speeds and the ability to pause and resume, users can interactively learn how these algorithms work. Whether you're a student or just interested in the fascinating world of algorithms, this project offers an engaging and educational experience